Revise logo to pretty much look like the original
Ultimately, I just have to admit, none of my attempts to redraw the logo
by just eyeballing it and trusting the grid ever really sized up to the
original, free-drawn version.

Sure, the coordinates don't match up to make it properly symmetrical or
evenly-spaced, but that doesn't matter, visually, as closure does the
work so that, ultimately, the uneven geometry looks more realistic.

Anyway, this version still fits all keypoints to even integer
coordinates, so that key nitpick is fixed, along with all the technical
improvements regarding transforms and
the like. It also, along in the process of fitting all the rules to this
grid, I feel smoothes out the few points on the original that I felt
were slightly awkward (like the *really* wonky top curve, which I feel
does not lose something in being more closely-centered).

Doing a straight comparison to the original, I feel this isn't quite
there yet, but I'm getting close.