Just get as close to the original logo as possible
You know what? The original, free-drawn logo was pretty much perfect,
and there's not really anything to be gained, visually, in trying to
correct its aesthetic quirks. Every time I tried to fit a relation more
closely to the grid in order to make its geometry more "pure", I would
just do a blind test and found that the *less* logical and pure version
just looked more impressive.

As a result, this commit just uses the 128x128 grid to come as close to
the geometry of the original logo as possible, sticking to the 64x64 or
32x32 grid only when it doesn't substantially change the form of the
curves.

(The most significant change I refused to budge on is that the curve in
the middle truly straightens vertically at the inflexion point.)